course_api_scraping/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pymongo==4.10.1
python-dotenv==1.0.1